optimize float

+++ software/imu.c
@@ -63,7 +63,7 @@ void imu_get_data(float *acc_angle_deg, float *gyro_angular)
 
     // calculate angle from accelerometer data
     //*acc_angle_deg = -atan2f(acc_z, -acc_x) * (180/M_PI); 
-    *acc_angle_deg = - atan2f(acc_z, sqrtf(acc_y * acc_y + acc_x * acc_x)) * (180/M_PI);
+    *acc_angle_deg = -atan2f(acc_z, sqrtf(acc_y * acc_y + acc_x * acc_x)) * (float)(180/M_PI);
 
     int16_t gyro_x = gyro[0];
     int16_t gyro_y = gyro[1];
